pyproject.toml
[tool.poetry]
name = "bounty_hunter_mw2"
version = "0.1.0"
description = "A discord bot for tracking and reporting cheaters. "
authors = ["Dylan Garrett"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
discord = "^2.1.0"
sqlalchemy = {extras = ["aiosqlite"], version = "^2.0.0"}
aiosqlite = "^0.18.0"
aiohttp = "^3.8.3"
python-dotenv = "^0.21.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
